Copy a case project from one place, and name the exit wait
What the review found, and what the code now does.

The copy of a test project lived in four places, three of which skipped the
canonicalization the fourth documents, and the one that had it was filed
under the completion harness that eight unrelated files then imported. It is
`case_project` now, and the four copies are one.

The exit wait was the literal 500 milliseconds at every one of the seventeen
`shutdown` calls, so `shutdown` takes it from `EXIT_TIMEOUT` and takes no
argument. `RESPONSE_TIMEOUT` and `PASS_TIMEOUT` join it at the top of the
impl, where the constants are read together.

Reading what the reader thread has taken in asks for a shared borrow, so
`pop_message`, `take_response`, `wait_for_response` and `expect_response`
take one. A panic message built with `format!` is built on the path that
panics.

The main loop takes a pass's result in above the read that blocks, which is
what the test client's wait for the program is written against; the loop now
says so where a change to it would be made.
