+++ src/tests/test_lsp/completion_harness.rs
@@ -2,42 +2,14 @@
 // over a private copy of a test project and sends completion /
 // resolve requests against it.
 
+use super::case_project::setup_test_env;
 use super::lsp_client::LspClient;
-use crate::tests::test_util::copy_dir_recursive;
 use serde_json::{json, Value};
 use std::{
     path::{Path, PathBuf},
     time::Duration,
 };
 use tempfile::TempDir;
-
-/// The directory holding the LSP test projects, one subdirectory per
-/// project, named as the tests name it.
-fn get_test_cases_dir() -> PathBuf {
-    let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    path.push("src/tests/test_lsp/cases");
-    path
-}
-
-/// Copy the test project `project_name` into a temporary directory of its
-/// own, so tests that build and edit it can run in parallel.
-///
-/// # Returns
-/// The guard whose drop deletes the copy, and the canonicalized path of
-/// the copied project. Canonicalizing resolves the symlinks a temporary
-/// directory sits behind (`/tmp` -> `/private/tmp` on macOS), so the path
-/// matches the root URI the server is initialized with and the URIs it
-/// publishes under.
-pub fn setup_test_env(project_name: &str) -> (TempDir, PathBuf) {
-    let temp_dir = TempDir::new().expect("Failed to create temp directory");
-    let test_case_src = get_test_cases_dir().join(project_name);
-    let test_case_dst = temp_dir.path().join(project_name);
-    copy_dir_recursive(&test_case_src, &test_case_dst).expect("Failed to copy test case");
-    let test_case_dst = test_case_dst
-        .canonicalize()
-        .expect("Failed to canonicalize test case path");
-    (temp_dir, test_case_dst)
-}
 
 /// Look up the `sortText` of the completion item whose `label` is `label`.
 pub fn find_sort_text(items: &[Value], label: &str) -> Option<String> {
@@ -101,7 +73,7 @@ impl LspCompletionCtx {
         for f in files {
             client
                 .open_document(Path::new(f))
-                .expect(&format!("Failed to open {}", f));
+                .unwrap_or_else(|_| panic!("Failed to open {}", f));
         }
         let trigger_file = files.last().unwrap();
         client.save_and_wait_for_the_program(Path::new(trigger_file));
@@ -165,9 +137,7 @@ impl LspCompletionCtx {
 
     /// Shut the server down and assert its reader thread saw no errors.
     pub fn shutdown(mut self) {
-        self.client
-            .shutdown(Duration::from_millis(500))
-            .expect("Failed to shutdown LSP");
+        self.client.shutdown().expect("Failed to shutdown LSP");
         self.client
             .verify_no_protocol_error()
             .expect("Reader thread should not have errors");
