+++ src/tests/test_lsp/test_goto_definition.rs
@@ -4,7 +4,7 @@
 
 #[cfg(test)]
 mod tests {
-    use super::super::completion_harness::setup_test_env;
+    use super::super::case_project::setup_test_env;
     use super::super::lsp_client::LspClient;
     use serde_json::{json, Value};
     use std::{
@@ -37,7 +37,7 @@ mod tests {
             for f in files {
                 client
                     .open_document(Path::new(f))
-                    .expect(&format!("Failed to open {}", f));
+                    .unwrap_or_else(|_| panic!("Failed to open {}", f));
             }
             let trigger_file = files.last().unwrap();
             client.save_and_wait_for_the_program(Path::new(trigger_file));
@@ -76,9 +76,7 @@ mod tests {
 
         /// Shut the server down, and fail the test if its reader thread met a protocol error.
         fn shutdown(mut self) {
-            self.client
-                .shutdown(Duration::from_millis(500))
-                .expect("Failed to shutdown LSP");
+            self.client.shutdown().expect("Failed to shutdown LSP");
             self.client
                 .verify_no_protocol_error()
                 .expect("Reader thread should not have errors");
@@ -87,7 +85,8 @@ mod tests {
 
     /// Read the substring of `file` covered by an LSP range.
     fn read_text_at_range(file: &Path, range: &Value) -> String {
-        let content = fs::read_to_string(file).expect(&format!("Failed to read file: {:?}", file));
+        let content =
+            fs::read_to_string(file).unwrap_or_else(|_| panic!("Failed to read file: {:?}", file));
         let lines: Vec<&str> = content.lines().collect();
         let sl = range["start"]["line"].as_u64().unwrap() as usize;
         let sc = range["start"]["character"].as_u64().unwrap() as usize;
