+++ src/tests/test_lsp/test_request_handling.rs
@@ -6,7 +6,7 @@
 
 #[cfg(test)]
 mod tests {
-    use super::super::completion_harness::setup_test_env;
+    use super::super::case_project::setup_test_env;
     use super::super::lsp_client::LspClient;
     use serde_json::json;
     use std::path::{Path, PathBuf};
@@ -97,9 +97,7 @@ mod tests {
         // the message above sends no progress, and the wait times out.
         client.save_and_wait_for_the_program(lib_fix);
 
-        client
-            .shutdown(Duration::from_millis(500))
-            .expect("Failed to shutdown LSP");
+        client.shutdown().expect("Failed to shutdown LSP");
         client
             .verify_no_protocol_error()
             .expect("Reader thread should not have errors");
@@ -187,9 +185,7 @@ mod tests {
             decodable
         );
 
-        client
-            .shutdown(Duration::from_millis(500))
-            .expect("Failed to shutdown LSP");
+        client.shutdown().expect("Failed to shutdown LSP");
         client
             .verify_no_protocol_error()
             .expect("Reader thread should not have errors");
@@ -220,9 +216,7 @@ mod tests {
             "a uri naming no file on disk is expected to be answered with no symbols"
         );
 
-        client
-            .shutdown(Duration::from_millis(500))
-            .expect("Failed to shutdown LSP");
+        client.shutdown().expect("Failed to shutdown LSP");
         client
             .verify_no_protocol_error()
             .expect("Reader thread should not have errors");
